+++ motto-cgss.sln
@@ -5,9 +5,7 @@ VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "motto-cgss-core", "motto-cgss-core\motto-cgss-core.csproj", "{6CBA9CC6-6CEF-4E26-AFEA-5F0C645E358D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cgss2motto", "tools\cgss2motto\cgss2motto.csproj", "{825A9022-57D1-4D01-8DE9-A83AD9A4BBB2}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "osu2motto", "tools\osu2motto\osu2motto.csproj", "{E9F0DCD0-24A0-4197-9F03-D3CE934CFEDA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "motto-conv", "motto-conv\motto-conv.csproj", "{7357864E-D035-4419-8967-26A2EEDF89A6}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,14 +17,10 @@ Global
 		{6CBA9CC6-6CEF-4E26-AFEA-5F0C645E358D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6CBA9CC6-6CEF-4E26-AFEA-5F0C645E358D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6CBA9CC6-6CEF-4E26-AFEA-5F0C645E358D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{825A9022-57D1-4D01-8DE9-A83AD9A4BBB2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{825A9022-57D1-4D01-8DE9-A83AD9A4BBB2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{825A9022-57D1-4D01-8DE9-A83AD9A4BBB2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{825A9022-57D1-4D01-8DE9-A83AD9A4BBB2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E9F0DCD0-24A0-4197-9F03-D3CE934CFEDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E9F0DCD0-24A0-4197-9F03-D3CE934CFEDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E9F0DCD0-24A0-4197-9F03-D3CE934CFEDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E9F0DCD0-24A0-4197-9F03-D3CE934CFEDA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7357864E-D035-4419-8967-26A2EEDF89A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7357864E-D035-4419-8967-26A2EEDF89A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7357864E-D035-4419-8967-26A2EEDF89A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7357864E-D035-4419-8967-26A2EEDF89A6}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
