+++ motto-conv/Osu2Motto.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace motto_conv
+{
+    public static class Osu2Motto
+    {
+        public static string Convert(string[] lines, double bpm)
+        {
+            return "";
+        }
+    }
+}
